refactor(form/imputs): fix issues and improve PT-input custom rendring
This will fix an issue with lacking support for custom inlineBlock rendring, and streamline all render components to have schemaType as prop.
Also streamlined how we render the custom or default component. Props (like .children) can now be manipulated with the .renderDefault prop, as opposed to just return static content